filter out unconvertible suggestions
