+++ src/features/suggestions/search/autoSuggestion.ts
@@ -2,20 +2,11 @@ import _ from 'lodash';
 import { Suggestion } from "../Suggestion";
 import baseOn from './calculateServing';
 import { PredefinedSuggestion } from './search';
-import { Unit, parseAmount, unitOf, isMeasure, measureOf } from "../parser/amount";
-
-function isConvertible(fromUnit: Unit, suggestion: PredefinedSuggestion) {
-  const isConvertibleTo = _.partial(isMeasure, _, measureOf(fromUnit));
-  const toUnit = parseAmount(suggestion.amount).unit;
-  return isConvertibleTo(toUnit);
-}
 
 export function generateAutoSuggestion(autoCompletion: Suggestion, suggestions: PredefinedSuggestion[]) {
   if (_.size(suggestions) === 0) return null;
 
-  const isConvertibleFromAutoCompletion = _.partial(isConvertible, unitOf(autoCompletion.amount || ""));
-  const bestMatch = _.head(_.filter(suggestions, isConvertibleFromAutoCompletion))
-    || suggestions[0];
+  const bestMatch = suggestions[0];
   return createAutoSuggestion(autoCompletion, bestMatch);
 }
